Kill prototools/test_schemas
